+++ backend/windmill-common/src/variables.rs
@@ -6,12 +6,11 @@
  * LICENSE-AGPL for a copy of the license.
  */
 
-use crate::error::Result;
+use crate::error;
 use crate::{worker::WORKER_GROUP, BASE_URL, DB};
 use chrono::{SecondsFormat, Utc};
 use magic_crypt::{MagicCrypt256, MagicCryptError, MagicCryptTrait};
 use serde::{Deserialize, Serialize};
-use crate::error;
 
 lazy_static::lazy_static! {
     pub static ref SECRET_SALT: Option<String> = std::env::var("SECRET_SALT").ok();
@@ -134,7 +133,7 @@ pub async fn get_secret_value_as_admin(
         let value = variable.value;
         if !value.is_empty() {
             let mc = build_crypt(db, w_id).await?;
-            decrypt_value_with_mc(value, mc).await?
+            decrypt(&mc, value)?
         } else {
             "".to_string()
         }
@@ -143,16 +142,6 @@ pub async fn get_secret_value_as_admin(
     };
 
     Ok(r)
-}
-
-pub async fn decrypt_value_with_mc(value: String, mc: MagicCrypt256) -> Result<String> {
-    mc.decrypt_base64_to_string(value).map_err(|e| match e {
-        MagicCryptError::DecryptError(_) => crate::error::Error::InternalErr(
-            "Could not decrypt value. The value may have been encrypted with a different key."
-                .to_string(),
-        ),
-        _ => crate::error::Error::InternalErr(e.to_string()),
-    })
 }
 
 pub fn encrypt(mc: &MagicCrypt256, value: &str) -> String {