+++ frontend/src/lib/components/RunForm.svelte
@@ -18,6 +18,7 @@
 	import RunFormAdvancedPopup from './RunFormAdvancedPopup.svelte'
 	import { page } from '$app/stores'
 	import { replaceState } from '$app/navigation'
+	import JsonInputs from '$lib/components/JsonInputs.svelte'
 
 	export let runnable:
 		| {
@@ -54,9 +55,12 @@
 	export let overrideTag: string | undefined
 
 	export let args: Record<string, any> = {}
+	export let jsonView = false
 
 	let reloadArgs = 0
 	let blockPopupOpen = false
+	let jsonEditor: JsonInputs | undefined = undefined
+	let schemaHeight = 0
 
 	export async function setArgs(nargs: Record<string, any>) {
 		args = nargs
@@ -88,6 +92,10 @@
 		} catch (e) {
 			console.error('Impossible to set hash in args', e)
 		}
+	}
+
+	export function setCode(code: string) {
+		jsonEditor?.setCode(code)
 	}
 </script>
 
@@ -137,7 +145,7 @@
 	{#if topButton}
 		<Button
 			btnClasses="!px-6 !py-1 w-full"
-			disabled={!isValid}
+			disabled={!isValid || jsonView}
 			on:click={() => runAction(undefined, args, invisible_to_owner, overrideTag)}
 		>
 			{buttonText}
@@ -147,22 +155,37 @@
 		<div class="my-2" />
 		{#if !runnable.schema.properties || Object.keys(runnable.schema.properties).length === 0}
 			<div class="text-sm py-4 italic">No arguments</div>
+		{:else if jsonView}
+			<div class="py-2" style="height: {schemaHeight}px" data-schema-picker>
+				<JsonInputs
+					bind:this={jsonEditor}
+					on:select={(e) => {
+						if (e.detail) {
+							args = e.detail
+						}
+					}}
+					updateOnBlur={false}
+					placeholder={`Write args as JSON.<br/><br/>Example:<br/><br/>{<br/>&nbsp;&nbsp;"foo": "12"<br/>}`}
+				/>
+			</div>
 		{:else}
 			{#key reloadArgs}
-				<SchemaForm
-					helperScript={runnable.hash
-						? {
-								type: 'hash',
-								hash: runnable.hash
-						  }
-						: undefined}
-					prettifyHeader
-					{noVariablePicker}
-					{autofocus}
-					schema={runnable.schema}
-					bind:isValid
-					bind:args
-				/>
+				<div bind:clientHeight={schemaHeight}>
+					<SchemaForm
+						helperScript={runnable.hash
+							? {
+									type: 'hash',
+									hash: runnable.hash
+							  }
+							: undefined}
+						prettifyHeader
+						{noVariablePicker}
+						{autofocus}
+						schema={runnable.schema}
+						bind:isValid
+						bind:args
+					/>
+				</div>
 			{/key}
 		{/if}
 	{:else}
@@ -176,7 +199,7 @@
 					{loading}
 					color="dark"
 					btnClasses="!px-6 !py-1 !h-8 inline-flex gap-2"
-					disabled={!isValid}
+					disabled={!isValid || jsonView}
 					on:click={() => runAction(scheduledForStr, args, invisible_to_owner, overrideTag)}
 					shortCut={{ Icon: CornerDownLeft, hide: !viewKeybinding }}
 				>
@@ -230,7 +253,7 @@
 	{:else if !topButton}
 		<Button
 			btnClasses="!px-6 !py-1 w-full"
-			disabled={!isValid}
+			disabled={!isValid || jsonView}
 			on:click={() => runAction(undefined, args, invisible_to_owner, overrideTag)}
 			shortCut={{ Icon: CornerDownLeft, hide: !viewKeybinding }}
 		>